+++ RELEASE_NOTES.md
@@ -7,12 +7,14 @@ Release Notes
 * Functions: Support for 64 bits.
 * Storage: Add support for tables
 * Event Grid: Ensure destination Queues are created as a dependency
+* Event Grid: Add ServiceBus Queue and Topic as supported destinations
 * Web App: Disables the automatic addition of the logging site extension when `docker_image` is used
 * ServiceBus: update namespace validation rules to follow [Microsoft documentation](https://docs.microsoft.com/en-us/azure/azure-resource-manager/management/resource-name-rules#microsoftservicebus)
 
 * Framework: Updated DeterministicGuid for RFC 4122 compatibility
 * Framework: Add support for NET5, upgrade to F#5.
 * Framework: Simplify Event Grid builder
+
 
 ## 1.4.0
 * Bing Search: Support for Bing Search (migrated from Cognitive Services).